🔥 Delete DateUnit.java
DateUnit is used in the previous version of `pointInTimeJoin( )` method. However, due to the complexity of the implementation, I just use nested for loop with time complexity of O(n^3) for the method. Even it's becoming more time-consuming when n is getting larger, i.e the dataset becomes larger, it's more readable and easy to implement.